% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjSurv.fn.R
\name{AdjSurv.fn}
\alias{AdjSurv.fn}
\title{Active Tag-Life-Adjusted Survival Modeling Fitting}
\usage{
AdjSurv.fn(
  taghist.file,
  taghist.format = "atlas",
  taglife.file = NULL,
  taglife.model = NULL,
  num.release = 1,
  num.bootstrap.in = 250,
  adjust.cjs = T,
  rounding = 4,
  plot.taglife = T,
  study.type = "single"
)
}
\arguments{
\item{taghist.file}{Required. Dataframe or a filename of .csv file with detection histories}

\item{taghist.format}{Required. Format of tag detection histories
\describe{
\item{"atlas"}{format based on Program ATLAS input files. Eight columns. Each tag has one line per possible detection site.}
\item{"flat"}{format has one line per tag: tag.code, activation date, release date, and
one column per detection site filled with first detection times at that site}
}}

\item{taglife.file}{Optional. Name of .csv file with taglife times in 1st column.  other columns ignored.  Header expected.}

\item{taglife.model}{fc_obj. Failure time model object. (default = NULL).  If no fc_obj is provided, function will try models available in failCompare and select best fit}

\item{num.release}{(default = 1) If more than 1 release group, 1st column will be added to flat format file to  denote group name}

\item{num.bootstrap.in}{(default = 1000) Number of resample iterations to estimate additional variance on survival estimates}

\item{adjust.cjs}{(T|F) (default = T) adjust CJS estimates be adjusted for estimated tag-life}

\item{rounding}{(default = 4)  Number of decimal places on estimate}

\item{plot.taglife}{(T|F) (default = T) Plot the estimated tag-life curve}

\item{study.type}{analyze ("single"|"paired") releases.}
}
\value{
Returns a list "out" with (if provided):
\describe{
\item{taghist}{tagfile filename}
\item{unadjusted.cjs}{table of unadjusted survival estimates}
\item{tagfile}{tag life filename}
\item{taglife.model}{taglife model name}
\item{mean.tag.pLive}{expect proportion tags active to each site}
\item{adjust.cjs}{adjusted survival estimates}
}
}
\description{
This function analyzes release-recapture data from a study that uses active-tag technology.
It estimates survival and detection probabilities and, if provided with tag-life data, the average probability
a tag is active at each detection site. These estimates are then used to adjust estimated survival for
potential tag-failure that may otherwise be interpreted as mortality. Methods are documented in
\href{http://www.cbr.washington.edu/sites/default/files/manuals/ATLAS_1_4_Manual_0.pdf}{Program ATLAS 1.4:Active Tag Life Adjusted Survival}.
}
