% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs.paired.fn.R
\name{cjs.paired.fn}
\alias{cjs.paired.fn}
\title{Summarizes detection histories for a paired release study provides
Cormack-Jolly-Seber estimates with or without adjustment of potential tag-failure
INPUT}
\usage{
cjs.paired.fn(
  detects.in,
  L.in = NULL,
  seeds.in = NULL,
  se.out = F,
  d = NULL,
  common.start.in = 0,
  parscale.in = 1e-04,
  reltol.in = 1e-30,
  maxit.in = 2e+06
)
}
\arguments{
\item{detects.in}{individual vectors of c(0,1,2) detection histories in wide ATlAS format, w/o rel.group or bin columns}

\item{L.in}{Vector of mean probabilities tag is working given detection at each site. If NULL, function will estimate the CJS parameters w/o tag failure (Li = 1)}

\item{seeds.in}{optional starting estimates for CJS model, includes additional p(censoring) per site at end parameters are Si, pi, li where i is 1:(total
number of sites-1). di (p(censoring)), is not estimated in this version}

\item{se.out}{(T/F) calculate se(cjs var) from variance matrix of parameters using Hessian?
OUTPUT
cjs.params: 2-column matrix with CJS estimates and standard errors (if estimated).

hessian:  variable to indicate if hessian matrix to be estimated or not
This is used to speed up estimation within the program}

\item{d}{proportion of detected tags censored at a location}

\item{common.start.in}{(0 to 2*(num.period-1)) indicates location common parameters start.0=all unique}

\item{parscale.in}{precision of estimates using optim.  Default is 0.0001}

\item{reltol.in}{relative convergence tolerance in optim.  Default is 2e+06 for initial estimate, 0.00001 when estimating SE.}

\item{maxit.in}{maximum number of iterations using optim. Default is 1e-30 for initial estimate, 200 when estimating SE.}

\item{detects}{a list of lists (individual detect.in): upper and lower release groups}
}
\value{
Returns a two-column matrix with CJS estimates and standard errors
(if estimated)
}
\description{
CJS.paired.fn summarizes the individual detection histories and estimates seeds for each releases Cormack-Jolly-Seber (CJS) likelhood function. It is based
on Skalski et al. (1998). This likelihood is also used to adjust survival for estimated tag life as described in Townsend et al. (2006).  If no tag-life
data provided, only the unadjusted CJS survival will be returned.
}
