% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic_tt2site.fn.R
\name{harmonic.tt2site.fn}
\alias{harmonic.tt2site.fn}
\title{Estimates travel time harmonic mean}
\usage{
harmonic.tt2site.fn(data.in, num.period, site.names)
}
\arguments{
\item{data.in}{Detection history and detection times (flat format)}

\item{num.period}{Number of detection sites}

\item{site.names}{Vector of site names}
}
\value{
This function returns a list with:
\describe{
\item{harmonic.tt}{harmonic mean travel time}
\item{tt.matrix}{individual travel time, release to first detection time for each site (days)}
\item{activetime.matrix}{individual time tag active, time of tag activation to first detection for each site (days)}
}
}
\description{
This function estimates the travel time harmonic mean.
}
