% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microtag_taglife}
\alias{microtag_taglife}
\title{taglife data with a shorter expected life}
\format{
Date set with 50 observations, 1 column
\describe{
\item{time.active}{total time (days) tag observed operating}
}
}
\source{
created to illustrate taglife affect on survvial estimation
}
\usage{
microtag_taglife
}
\description{
taglife data with a shorter expected life
}
\keyword{datasets}
