% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired.cjs.lik.R
\name{paired.cjs.lik}
\alias{paired.cjs.lik}
\title{A likelihood function for multiple unadjusted- or adjusted-for-taglife Cormack-Jolly-Seber (CJS) estimates
that have common parameters (S, p, l)}
\usage{
paired.cjs.lik(
  params.in,
  counts.in,
  num.period,
  use.hist,
  L.in = NULL,
  d.in = NULL,
  common.start = NULL,
  fix.1 = NULL
)
}
\arguments{
\item{params.in}{List of likelihood MLE estimators (S,p,l) in a single vector (1 each for upper, lower release)}

\item{counts.in}{List of Vectors of summed detections by history (names are history)}

\item{num.period}{Number of detection periods in history}

\item{use.hist}{List of Matrices of unique detection histories}

\item{L.in}{List of probability that a tag is working at a detection site at a given time. Set to 1 if active tags not indicated.
p(tag working) is cumulative to each site, not a multiplicate from site to site}

\item{d.in}{List of probability a tag is censored given that it's detected}

\item{common.start}{numeric indicating point survival/detection p's are equated}

\item{fix.1}{vector of estimates greater than 1 that will be fixed to 1}
}
\value{
This function returns a negative log-likelihood
}
\description{
Based on the likelihood function for unadjusted- or adjusted-for-taglife
Cormack-Jolly-Seber (CJS) estimates and Skalski et al. (1998)
Adjustment to survival for estimated tag life as described in Townsend et al.
(2006).  User input indicates which parameters should be set to the same value. This
function is set above individual CJS likelihood calls
}
