% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paired.rel.practice}
\alias{paired.rel.practice}
\title{paired release study data}
\format{
Date frame with 9996 observations for first release, 8000 obs for second release , 8 columns
\describe{
\item{release designation}{name}
\item{bin}{replicate indicator from same release site}
\item{tag id}{code}
\item{tag activation time}{date-time}
\item{release time}{date-time}
\item{detection site}{date-time}
\item{detection type}{0,1, or 2}
\item{detection time at that site}{date-time}
}
}
\usage{
paired.rel.practice
}
\description{
paired release study data
}
\keyword{datasets}
