% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs2.fn.R
\name{vs2.fn}
\alias{vs2.fn}
\title{Estimates a ratio of 2 survival estimates w/standard errors}
\usage{
vs2.fn(s1, s2, se1, se2, se.in = T, se.output = T)
}
\arguments{
\item{s1}{survival upper reach}

\item{s2}{survival lower reach}

\item{se1}{survival upper reach standard error}

\item{se2}{survival lower reach standard error}

\item{se.in}{standard errors or variances being used for inputs, default = SE}

\item{se.output}{output of survival ratio point estimate with standard error or variance, default = SE}
}
\value{
This function returns a vector with:
\describe{
\item{point estimate with standard error or variance}{s1 = upper release survival, s2=lower release survival, se's are corresponding standard errors}
}
}
\description{
This function uses the point and standard estimates to get the adjust survival and standard error. default is no covariance between survivals
}
